Add PictureSetPath property
Added a PictureSetPath property.  There's already a LoadImage and LoadImageFromPath property but these were not intuitive for me.  I want all Picture related properties to appear near each other in the Property Pages and to be named similarly.  Now this matches PictureSetWidth and PictureSetHeight, etc...

PictureSetPath calls LoadImageFromPath when set.